requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.0.0

# EEG processing & signal analysis
PyWavelets>=1.1.0
mne>=1.0.0

# Machine learning
xgboost>=1.5.0

# Deep learning (optional - for EEGNet)
# torch>=1.9.0
# torchvision>=0.10.0

# Visualization & GUI
streamlit>=1.20.0
streamlit-autorefresh>=0.0.1
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pygame>=2.0.0

# Audio processing (optional - for actual playback)
# pygame>=2.0.0

# Utilities
pyyaml>=6.0
h5py>=3.0.0
tqdm>=4.0.0
reportlab>=3.6.0
psutil>=5.8.0  # System and process utilities (for runtime metrics)

# Testing
pytest>=7.0.0

# Advanced Features (NEW)
openpyxl>=3.0.0          # Excel export
scikit-learn>=1.0.0      # Already have, but ensure latest
statsmodels>=0.13.0      # Statistical analysis
pingouin>=0.5.0          # Additional statistical tests